+++ tests/assets.py
@@ -1,0 +1,54 @@
+CONTENT_TO_SORT = """
+Some text
+# Main title
+
+## 2020-01-03 - Some other title
+Some other text
+
+## 2020-01-01 - Some title
+Some text
+## Some non-timestamp title
+non-timestamp text
+## 2020-01-02 - Some other title
+Some other text
+"""
+
+
+SORTED_CONTENT_REVERSED = """Some text
+# Main title
+
+
+## Some non-timestamp title
+non-timestamp text
+
+
+## 2020-01-03 - Some other title
+Some other text
+
+
+## 2020-01-02 - Some other title
+Some other text
+
+
+## 2020-01-01 - Some title
+Some text"""
+
+
+SORTED_CONTENT = """Some text
+# Main title
+
+
+## Some non-timestamp title
+non-timestamp text
+
+
+## 2020-01-01 - Some title
+Some text
+
+
+## 2020-01-02 - Some other title
+Some other text
+
+
+## 2020-01-03 - Some other title
+Some other text"""